backend/codecompletion/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
bcrypt==3.2.2
boto3==1.24.43
botocore==1.27.43
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
datefinder==0.7.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.51.0
django-bootstrap3==21.2
django-braces==1.15.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-oauth-toolkit==2.0.0
django-rest-swagger==2.2.0
django-storages==1.12.3
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-social-oauth2==1.2.1
drf-yasg==1.20.0
ecdsa==0.17.0
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.2
MarkupSafe==2.1.1
misaka==2.1.1
mysqlclient==2.1.0
numpy==1.21.6
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.1
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
unicodecsv==0.14.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
wincertstore==0.2
wrapt==1.14.0
